Refactor component folder structure to use subfolders for related components
